+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Update to MLflow 2.7.1
 - Improve `table_exists()` in `example_merlion.py`
 - SQLAlchemy: Use server-side `now()` function for "autoincrement" columns
+- Use SQLAlchemy patches and polyfills from `cratedb-toolkit`
 
 ## 2023-09-12 0.1.1
 - Documentation: Improve "Container Usage" page